Add initial start screen
There is no meaningful default value for the output directory, but it is
the one option that should be set before people start using Gyre. This
is especially important for Flatpak builds as the empty path string will
map to a location within the Flatpak sandbox, thus downloading files to
an unreachable location.

We can't expect users to open the preferences before they start their
first download, so we have to prompt them for an output directory even
before that. That is what the new welcome screen/initial setup
accomplishes.