Replace Entry with ComboBox for adding communities
Afaik there are no hidden communities that would warrant an entry to
input the community name. So instead switch to a ComboBox for easier
selection. It also allows us to show the pretty community names instead
of the internally used ones.

Now I only need to find out if that changes Gyre's age rating as we
technically promote NSFW content.