Remove explicit usage of FFmpeg's file protocol
This is a remnant of CoubDownloader code and ensured that FFmpeg could
handle files with colons in the filename (FFmpeg assumes the usage of a
protocol in that case). Gyre is a lot more strict when it comes to
filenames, so this isn't an issue anymore.

Even without the "file:" prefix, FFmpeg still uses the file protocol to
handle input files. That's why the custom FFmpeg build flags stay the
same.