Clean up UI files
Glade is a great program to quickly design UIs, but it also has a bad
habit of bloating the files with unnecessary properties and placeholders.
Not to mention that you don't fully know what Glade might do to a file
when you hit save. Sometimes it adds just a bunch of new properties.
Sometimes it deletes widgets, because it isn't able to handle them
(e.g. Handy.ActionRows).

Gyre's design is basically set and smaller changes can still be made by
hand. So it's time to clean up the UI files and get rid of unnecessary
filler. Functionality-wise there aren't any obvious changes. In theory
the default widget focus for AddWindows gets removed, but that didn't
work before anyway.