Remove Gtk.FileChooserButton
The FileChooserButton has always been an unfortunate widget. It doesn't
support Gtk.FileChooserAction.SAVE, so I couldn't use it half of the
time anyway and the arbitrarily long list is an eyesore, especially if
it's long enough to exceed the screen. But there are also bigger issues,
which elevated this step from a distant goal to an immediate concern.

While the raised FileChooserDialog uses portals, when running Gyre as a
Flatpak, the dropdown list does not. So choosing an output directory
from the list led to coubs being saved in an inaccessible location. Not
good. On top of that the FileChooserButton is deprecated. Not in GTK3,
but in GTK4, which I plan to eventually use for Gyre.

Now Gyre uses a custom FileChooserButton. Nothing grand. Just a normal
button that will raise a FileChooserNative dialog when clicked and
replace its label with the currently selected file/folder and an image
to indicate whether the button handles files or folders.
The one drawback is that the button can't be included in UI files
anymore and needs to be added manually instead.