Update local build script
The most important change is that it runs "meson test". I didn't know
you had to do that explicitly. I assumed it would throw errors during
the configuration or build step automatically. Would've saved me quite a
bit of time for my Flathub submission.

Anyway, now the script is more suited for quick testing. The new name is
supposed to reflect that.