Add stream integrity check and repair
An old CoubDownloader feature that finally made its way into this
project. Corrupted streams are a rare sight nowadays, but you can
occasionally still run into them. The likeliest occurrence are old video
streams, which are still stored in a broken state, like they were before
Coub introduced watermarks to protect their "property". But corruption
can also be caused by a faulty transmission, Coub removing the file from
their servers while it's being downloaded or because the file was
(accidentally) broken to begin with.

As long as the corruption is caused by the old storage method, we can
fix it quite easily. Everything else is treated as an error.

The one drawback to these additional checks is that we have to compile
the custom FFmpeg build with more options to support H264 and MP3
decoding. But overall a minor trade-off.