Don't disconnect signals on container error
There's no point keeping track of an item's status once it threw an
error. Either the item failed to parse and provides no links or
something crashed the whole processing function.
The previous behavior was to freeze the notify signals, so that they
don't fire in the first place. This was an easy solution and worked well
for invalid containers (e.g. when a wrong ID was specified), as the user
would have to update the item's info to fix the problem, which would
destroy the old frozen container anyway.
However, invalid items are not always the cause for errors. For example,
connection issues could lead to a container not being parsed. In such a
case simply restarting the download can solve the issue, but the item's
signals were already frozen, so the progress would never update. The
only solution was to re-add the item.

That's clearly not desirable. It should be possible to retry erroneous
containers and still see the progress. So instead of freezing the notify
signals, a new custom setter makes sure that an error status doesn't
get updated, unless said status marks the beginning of a new download
attempt.