Add periodic downloads
Another feature that I was dearly missing from the original CLI script
(or rather my personal wrapper for it). The option can be found
in the "Download" tab.

As for the implementation, I decided to keep the worker thread alive and
simply loop inside the processing function. That means a periodic
download behaves just like a never ending single download, locking the
interface in certain places (e.g. no new items, no profile import). The
user has to stop the download manually, before they can edit the input
item list again.

The drawback is that all of this bloats the processing function more and
more. I also had to introduce another makeshift global variable to
signal the processing function to stop, while waiting for the next
download iteration to start. Overall pretty messy. But fixing that is a
task for another day.