Move duplicate/archive check logic to own module
The goal is to move these checks out of the main processing function and
instead make each container responsible for discarding IDs as they are
being parsed. This is the first of many steps to remove the special
treatment of the parsing process and to make each container independent
from start to finish (right now parsing has to be done in sequential
order).

As an additional note, I did test the viability of databases, as an
alternative to reading text files into giant sets, but at least the
sqlite3 module doesn't provide the necessary performance.
It would've also made it very difficult to use custom archive paths with
the Flatpak version, as SQLite requires write access to the database
directory.