Move coub download logic to containers
As announced this commit changes how Gyre handles the parsing and
download process. Until now the main processing function was responsible
to coordinate these steps. It was a simple solution. One that has been
part of CoubDownloader (and now Gyre) for a long time.
There were, however, many downsides to this. The most obvious problem
for a user was the sequential parsing process. With many input items it
could take minutes of just fetching IDs, before the actual download
process started. But it also led to some unnecessary complexity in the
GUI code (regarding item status feedback and general list protection).

With this commit each container is now a separate entity, responsible
for fetching and downloading the coubs they provide. It is pretty much a
copy of the original procedure, so a lot of the aforementioned
complexity is still in place, but I will address it in future commits.