Streamline container exception handling
On days like this, when Coub just refuses half of my connections for no
reason, it is best to work on exception handling.
As the try-blocks responsible for fetching Coub API JSON responses are
so small, basically anything that can go wrong in there indicates a
faulty connection. So let's just catch everything and blame it on Coub
or the user's network. Also stop reporting specific error messages, when
they are in reality only educated guesses. If we can't fetch the total
amount of pages from the API then it doesn't necessarily mean the user
specified an invalid container (e.g. non-existing tag). Just report that
we couldn't get the mandatory API info.

coub.py will get the same treatment, once I get to it.