Improve container error handling (again)
That also includes reverting some of the previous changes. The general
exception handling led to uncaught errors during the asyncio loop
shutdown, when the download gets cancelled by the user and potentially
displayed an APIResponseError message, when the items was actually
cancelled.
Additionally all items (incl. single coubs) now test for an error
message in the API response, which indicates nonexistence.