Unify cancellation process
Until now coubs, containers and the main processing function had to be
cancelled separately. A messy solution. The new approach looks towards
the utils module to provide the means to cancel all parts of the
download process at once via a decorator that works on both synchronous
and asynchronous functions.
Because this doesn't litter the code with dozens of if checks for global
variables, more functions/methods are now cancellable, which usually
leads to a faster cancellation of the entire process.

The main processing function was also slightly restructured to make use
of the CancelledErrors to stop repeated downloads.