Handle Coub server maintenance page
Occasionally the Coub API gets rather unstable, which leads to severe
connection issues. However, when a request can't be handled, instead of
returning an error code or a JSON with an error message, Coub redirects
to an apology website. This would lead to Gyre failing with an unknown
error, as the response couldn't be decoded as JSON.

Now we catch those decoding errors and treat them like a failed
connection attempt.